new AI movement algorithms

+++ Assets/_Scripts/Interfaces/BuffableInterfaces.cs
@@ -1,12 +1,9 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
 
 namespace BomberGame
 {
     public interface ILengthBuffable
     {
-        public void BuffLength(float length);
+        public void BuffLength(int length);
     }
     public interface IPierceBuffable
     {
